% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment3.R
\docType{package}
\name{alignment3}
\alias{alignment3}
\alias{alignment3-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
alignment3
A data package for alignment3.

Use \code{data(package='alignment3')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
