% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_filter_reference_doc_split.R
\name{clean_and_filter_reference_doc_split}
\alias{clean_and_filter_reference_doc_split}
\title{Clean and filter reference doc split

Clean block names mainly so that later regex works easier.
Filter out blocks that we don't need to match (and extract as DVs)}
\usage{
clean_and_filter_reference_doc_split(reference_doc_split)
}
\arguments{
\item{reference_doc_split}{}
}
\value{

}
\description{
Clean and filter reference doc split

Clean block names mainly so that later regex works easier.
Filter out blocks that we don't need to match (and extract as DVs)
}
\author{
Shir Dekel
}
