% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment3.R
\docType{data}
\name{data}
\alias{data}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{id}{}
\item{alignment}{}
\item{reliability_amount}{}
\item{phase}{}
\item{response_id}{}
\item{recorded_date}{}
\item{duration_in_seconds}{}
\item{business_edu}{}
\item{business_edu_other}{}
\item{business_exp}{}
\item{order_b}{}
\item{display}{}
\item{eval_align}{}
\item{order}{}
\item{product}{}
\item{project}{}
\item{ranking}{}
\item{allocation}{}
\item{confidence}{}
\item{rating}{}
\item{npv_knowledge}{}
\item{npv_amount}{}
\item{sample}{}
\item{npv_allocation_correlation}{}
\item{max_npv}{}
\item{min_npv}{}
\item{max_min_difference}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{alignment3}
}
