% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_masters.R
\name{filter_masters}
\alias{filter_masters}
\title{Filter Masters data}
\usage{
filter_masters(data_raw)
}
\arguments{
\item{data_raw}{}
}
\value{

}
\description{
Filter Masters data
}
\author{
Shir Dekel
}
