% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dv_id_match.R
\name{get_dv_id_match}
\alias{get_dv_id_match}
\title{Get dv id match}
\usage{
get_dv_id_match(content, dv_match_regex)
}
\arguments{
\item{content}{}

\item{dv_match_regex}{}
}
\value{

}
\description{
Get dv id match
}
\author{
Shir Dekel
}
