% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dv_id_name.R
\name{get_dv_id_name}
\alias{get_dv_id_name}
\title{Get DV ID name

Recode}
\usage{
get_dv_id_name(dv_id_match)
}
\arguments{
\item{dv_id_match}{}
}
\value{

}
\description{
Get DV ID name

Recode
}
\author{
Shir Dekel
}
