% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dv_id_name_replacement.R
\name{get_dv_id_name_replacement}
\alias{get_dv_id_name_replacement}
\title{Get DV ID name replacement

Prepare a named vector for subsequent recoding at \code{get_dv_id_name()}}
\usage{
get_dv_id_name_replacement(dv_match_pattern)
}
\value{

}
\description{
Get DV ID name replacement

Prepare a named vector for subsequent recoding at \code{get_dv_id_name()}
}
\author{
Shir Dekel
}
