% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dv_match.R
\name{get_dv_match}
\alias{get_dv_match}
\title{Get dv match variables

Used for subsequent renaming of column names using \code{dv_match$dv_pattern} and \code{dv_match$dv_replacement} in \code{rename_masters()}}
\usage{
get_dv_match(reference_doc_split)
}
\arguments{
\item{reference_doc_split}{}
}
\value{

}
\description{
Get dv match variables

Used for subsequent renaming of column names using \code{dv_match$dv_pattern} and \code{dv_match$dv_replacement} in \code{rename_masters()}
}
\author{
Shir Dekel
}
