% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dv_match_components.R
\name{get_dv_match_components}
\alias{get_dv_match_components}
\title{Get dv match components

Get table that ends up being \code{dv_match}, with the critical components \code{dv_pattern} and \code{dv_replacement}, and the elements that constructed them}
\usage{
get_dv_match_components(reference_doc_split_cleaned_and_filtered)
}
\arguments{
\item{reference_doc_split_cleaned_and_filtered}{}

\item{dv_match_pattern}{}
}
\value{

}
\description{
Get dv match components

Get table that ends up being \code{dv_match}, with the critical components \code{dv_pattern} and \code{dv_replacement}, and the elements that constructed them
}
\author{
Shir Dekel
}
