% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dv_match_pattern.R
\name{get_dv_match_pattern}
\alias{get_dv_match_pattern}
\title{Get DV match pattern

For matching DVs in the reference doc (to then find out the corresponding question ID)}
\usage{
get_dv_match_pattern()
}
\value{

}
\description{
Get DV match pattern

For matching DVs in the reference doc (to then find out the corresponding question ID)
}
\author{
Shir Dekel
}
