% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dv_match_regex.R
\name{get_dv_match_regex}
\alias{get_dv_match_regex}
\title{Get dv match regex

Generate regex that will be used to extract question IDs
The idea is to match a DV pattern and then go back and find the question ID associated with it (at the beginning of that block)}
\usage{
get_dv_match_regex(dv_match_pattern)
}
\arguments{
\item{dv_match_pattern}{}
}
\value{

}
\description{
Get dv match regex

Generate regex that will be used to extract question IDs
The idea is to match a DV pattern and then go back and find the question ID associated with it (at the beginning of that block)
}
\author{
Shir Dekel
}
