% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dv_pattern.R
\name{get_dv_pattern}
\alias{get_dv_pattern}
\title{Get DV pattern}
\usage{
get_dv_pattern(dv_id, dv_pattern_suffix)
}
\arguments{
\item{dv_id}{}

\item{dv_pattern_suffix}{}
}
\value{

}
\description{
Get DV pattern
}
\author{
Shir Dekel
}
