% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dv_replacement_masters.R
\name{get_dv_replacement_masters}
\alias{get_dv_replacement_masters}
\title{Get DV replacement for masters Experiment}
\usage{
get_dv_replacement_masters(
  block,
  dv_id_name,
  phase,
  product,
  dv_replacement_suffix
)
}
\arguments{
\item{product}{}
}
\value{

}
\description{
Get DV replacement for masters Experiment
}
\author{
Shir Dekel
}
