% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_npv_allocation_correlation.R
\name{get_npv_allocation_correlation}
\alias{get_npv_allocation_correlation}
\title{Get NPV x allocation correlation}
\usage{
get_npv_allocation_correlation(data_clean)
}
\arguments{
\item{data_clean}{}
}
\value{

}
\description{
Get NPV x allocation correlation
}
\author{
Shir Dekel
}
