% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_product.R
\name{get_product}
\alias{get_product}
\title{Get product variable

We need product data for each project for later analysis, because there was an error in the way the materials were entered into Qualtrics
This isn't otherwise available for each product, so we need to scrape the product names that exist for each block, and assign them the vector 1:5.

Uses \code{product_regex} from \code{get_product_regex()}}
\usage{
get_product(content, product_regex)
}
\arguments{
\item{content}{}

\item{product_regex}{}
}
\value{

}
\description{
Get product variable

We need product data for each project for later analysis, because there was an error in the way the materials were entered into Qualtrics
This isn't otherwise available for each product, so we need to scrape the product names that exist for each block, and assign them the vector 1:5.

Uses \code{product_regex} from \code{get_product_regex()}
}
\author{
Shir Dekel
}
