% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_product_regex.R
\name{get_product_regex}
\alias{get_product_regex}
\title{Generate regex to match product names

We need product data for each project for later analysis, because there was an error in the way the materials were entered into Qualtrics
This isn't otherwise available for each product, so we need to scrape the product names that exist for each block, and assign them the vector 1:5.}
\usage{
get_product_regex()
}
\value{

}
\description{
Generate regex to match product names

We need product data for each project for later analysis, because there was an error in the way the materials were entered into Qualtrics
This isn't otherwise available for each product, so we need to scrape the product names that exist for each block, and assign them the vector 1:5.
}
\author{
Shir Dekel
}
