% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_doc.R
\name{get_reference_doc}
\alias{get_reference_doc}
\title{Extract text from Qualtrics .docx export.

Needed because the first row of the dataframe doesn't contain information about within-subjects conditions.}
\usage{
get_reference_doc()
}
\value{

}
\description{
Extract text from Qualtrics .docx export.

Needed because the first row of the dataframe doesn't contain information about within-subjects conditions.
}
\author{
Shir Dekel
}
