% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_doc_split.R
\name{get_reference_doc_split}
\alias{get_reference_doc_split}
\title{Get reference document split

Split up the reference doc by blocks so that each one can be subsequently matched}
\usage{
get_reference_doc_split(reference_doc)
}
\value{

}
\description{
Get reference document split

Split up the reference doc by blocks so that each one can be subsequently matched
}
\author{
Shir Dekel
}
