% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_doc_split_regex.R
\name{get_reference_doc_split_regex}
\alias{get_reference_doc_split_regex}
\title{Get reference doc split regex
Split the text up for subsequent matching}
\usage{
get_reference_doc_split_regex()
}
\value{

}
\description{
Get reference doc split regex
Split the text up for subsequent matching
}
\author{
Shir Dekel
}
