% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_masters.R
\name{pivot_masters}
\alias{pivot_masters}
\title{Pivot Masters data}
\usage{
pivot_masters(masters_renamed)
}
\arguments{
\item{masters_renamed}{}
}
\value{

}
\description{
Pivot Masters data
}
\author{
Shir Dekel
}
