% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_business_exp.R
\name{recode_business_exp}
\alias{recode_business_exp}
\title{Recode business experience variable

Originally coded as \code{None}, \code{1}, \code{2}, \code{3} ...}
\usage{
recode_business_exp(business_exp)
}
\arguments{
\item{business_exp}{}
}
\value{

}
\description{
Recode business experience variable

Originally coded as \code{None}, \code{1}, \code{2}, \code{3} ...
}
\author{
Shir Dekel
}
