% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_project_masters.R
\name{recode_project_masters}
\alias{recode_project_masters}
\title{Recode masters experiment project numbers according to order variable}
\usage{
recode_project_masters(project, order)
}
\arguments{
\item{project}{}

\item{order}{}
}
\value{

}
\description{
Recode masters experiment project numbers according to order variable
}
\author{
Shir Dekel
}
