% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_reliability_masters.R
\name{recode_reliability_masters}
\alias{recode_reliability_masters}
\title{Recode reliability_amount for masters sample

Change high and low because originally coded as \code{variance}, which has opposite coding to \code{reliability_amount}. For instance, high variance implies low reliability.}
\usage{
recode_reliability_masters(reliability_amount)
}
\value{

}
\description{
Recode reliability_amount for masters sample

Change high and low because originally coded as \code{variance}, which has opposite coding to \code{reliability_amount}. For instance, high variance implies low reliability.
}
\author{
Shir Dekel
}
