% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_masters.R
\name{rename_masters}
\alias{rename_masters}
\title{Rename Masters data}
\usage{
rename_masters(masters_filtered, dv_match)
}
\arguments{
\item{dv_match}{}
}
\value{

}
\description{
Rename Masters data
}
\author{
Shir Dekel
}
